requirements.txt
pymongo>=4.7
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
python-telegram-bot>=21.0

